refactor: 도메인 패키지로 이동
- 도메인 관련 객체들을 도메인 패키지로 이동
