+++ src/main/java/calculator/domain/Operation.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.domain;
 
 import java.util.function.BiFunction;
