+++ src/test/java/calculator/domain/OperationTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
