+++ src/test/java/calculator/domain/OperationValidatorTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.domain;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
