+++ src/test/java/calculator/domain/StringSplitterTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
